Cast to size_t for signed value in assert
The `abs` function returns a signed `int`, is compared to an unsigned size_t
counter in ic3.

+++ src/ic3/build_prob/g2en_cnf.cc
@@ -72,9 +72,10 @@ int form_index(CUBE &C)
 {
   int result = 0;
   for (size_t i=0; i < C.size();i++)
-    {assert(abs(C[i]) == i+1);
-      if (C[i] > 0)  result |= 1 << i;
-    }
+  {
+    assert((size_t) abs(C[i]) == i+1);
+    if (C[i] > 0)  result |= 1 << i;
+  }
 
   return(result);
